Move to ~settings module
