+++ apps/test-app/app/~navigation.module.css
@@ -49,9 +49,3 @@
 .mainContent {
 	flex: 1;
 }
-
-.settingsDialogContent {
-	display: flex;
-	flex-direction: column;
-	gap: var(--stratakit-space-x4);
-}
