+++ apps/test-app/app/~navigation.tsx
@@ -5,31 +5,13 @@
 
 import * as React from "react";
 import { useHref, useLocation } from "react-router";
-import Dialog from "@mui/material/Dialog";
-import DialogContent from "@mui/material/DialogContent";
-import DialogTitle from "@mui/material/DialogTitle";
-import FormControl from "@mui/material/FormControl";
-import FormLabel from "@mui/material/FormLabel";
-import ToggleButton from "@mui/material/ToggleButton";
-import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
 import { Button, Divider, IconButton } from "@stratakit/bricks";
 import { Icon } from "@stratakit/foundations";
 import { unstable_NavigationRail as NavigationRail } from "@stratakit/structures";
-import {
-	isProduction,
-	useIsWideScreen,
-	useLocalStorage,
-	useSettingsStore,
-} from "./~utils.tsx";
+import { SettingsButton } from "./~settings.tsx";
+import { isProduction, useIsWideScreen, useLocalStorage } from "./~utils.tsx";
 
-import type { ExtractState } from "zustand";
-
-import svgComputer from "@stratakit/icons/computer.svg";
 import svgDocumentation from "@stratakit/icons/documentation.svg";
-import svgMoon from "@stratakit/icons/moon.svg";
-import svgSettings from "@stratakit/icons/settings.svg";
-import svgSun from "@stratakit/icons/sun.svg";
-import primitives from "internal/primitives.json" with { type: "json" };
 import styles from "./~navigation.module.css";
 import svgComponents from "./assets/components.svg";
 import svgIcons from "./assets/icons.svg";
@@ -217,92 +199,4 @@ interface RegularLinkProps
 /** Wrapper over `<a>` that resolves URLs using react-router.  */
 function RegularLink({ to, ...props }: RegularLinkProps) {
 	return <a href={useHref(to)} {...props} />;
-}
-
-// ----------------------------------------------------------------------------
-
-type SettingsState = ExtractState<typeof useSettingsStore>;
-type ColorSchemeSetting = SettingsState["colorScheme"];
-type AccentColor = SettingsState["accentColor"];
-
-function SettingsButton() {
-	const id = React.useId();
-	const [open, setOpen] = React.useState(false);
-
-	const colorScheme = useSettingsStore((state) => state.colorScheme);
-	const setColorScheme = useSettingsStore((state) => state.setColorScheme);
-
-	const accentColor = useSettingsStore((state) => state.accentColor);
-	const setAccentColor = useSettingsStore((state) => state.setAccentColor);
-	return (
-		<>
-			<NavigationRail.Button
-				label="Settings"
-				icon={svgSettings}
-				onClick={() => setOpen(true)}
-			/>
-			<Dialog open={open} onClose={() => setOpen(false)}>
-				<DialogTitle>Settings</DialogTitle>
-				<DialogContent className={styles.settingsDialogContent}>
-					<FormControl>
-						<FormLabel id={`${id}-color-scheme`}>Color scheme</FormLabel>
-						<ToggleButtonGroup
-							exclusive
-							value={colorScheme}
-							onChange={(_, value: ColorSchemeSetting | null) => {
-								setColorScheme(value === null ? "auto" : value);
-							}}
-							aria-labelledby={`${id}-color-scheme`}
-						>
-							<ToggleButton value="auto" label="Auto">
-								<Icon href={svgComputer} />
-							</ToggleButton>
-							<ToggleButton value="light" label="Light">
-								<Icon href={svgSun} />
-							</ToggleButton>
-							<ToggleButton value="dark" label="Dark">
-								<Icon href={svgMoon} />
-							</ToggleButton>
-						</ToggleButtonGroup>
-					</FormControl>
-					<FormControl>
-						<FormLabel id={`${id}-accent-color`}>Accent color</FormLabel>
-						<ToggleButtonGroup
-							exclusive
-							value={accentColor}
-							onChange={(_, value: AccentColor | null) => {
-								setAccentColor(value === null ? "aurora" : value);
-							}}
-							aria-labelledby={`${id}-accent-color`}
-						>
-							<ToggleButton value="aurora" label="Aurora">
-								<Icon
-									render={<ColorIcon />}
-									style={{
-										color: primitives.aurora[500],
-									}}
-								/>
-							</ToggleButton>
-							<ToggleButton value="cobalt" label="Cobalt">
-								<Icon
-									render={<ColorIcon />}
-									style={{
-										color: "oklch(53.32% 0.139 246.77)",
-									}}
-								/>
-							</ToggleButton>
-						</ToggleButtonGroup>
-					</FormControl>
-				</DialogContent>
-			</Dialog>
-		</>
-	);
-}
-
-function ColorIcon(props: React.ComponentProps<"svg">) {
-	return (
-		<svg {...props} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16">
-			<circle cx="8" cy="8" r="8" fill="currentColor" />
-		</svg>
-	);
 }
