+++ apps/test-app/app/~settings.module.css
@@ -1,0 +1,10 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+
+.dialogContent {
+	display: flex;
+	flex-direction: column;
+	gap: var(--stratakit-space-x4);
+}
